Validation tests for JavaScript in tape
